chore(github_releases): move utils from driver.go to util.go
